requirements.txt
torch
torchsde
torchvision
torchaudio
fastapi
pydantic
uvicorn
Pillow
